+++ sdb_ai/sd64/gplsrc/exepath.h
@@ -1,0 +1,39 @@
+/* EXEPATH.H
+ * Locate the directory holding the running executable.
+ * Copyright (c) 2026 Ladybridge Systems, All Rights Reserved
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 3, or (at your option)
+ * any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.
+ *
+ * START-HISTORY:
+ * 14 Aug 26 Windows port - new file
+ * END-HISTORY
+ *
+ * START-DESCRIPTION:
+ *
+ * END-DESCRIPTION
+ *
+ * START-CODE
+ */
+
+#ifndef __EXEPATH
+#define __EXEPATH
+
+/* Functions in exepath.c */
+
+bool exe_directory(char* buff, int buff_len);
+
+#endif
+
+/* END-CODE */